feat(infra): setup axios and react-query clients
