+++ src/lib/axios.ts
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+import { env } from "@/env";
+
+export const api = axios.create({
+  baseURL: env.VITE_API_URL,
+});
